refactor: rename session utils, remove useless modules
