+++ src/index.ts
@@ -2,7 +2,7 @@ import express from 'express'
 import http from 'http'
 import { Server } from 'socket.io'
 import { standardConfig } from './infrastructure/config'
-import { acceptCommands } from './infrastructure/adapters/websocket/acceptCommands'
+import { acceptCommands } from './infrastructure/acceptCommands'
 
 const app = express()
 const server = http.createServer(app)
